--- Scripts/NetworkManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c687836e5ac272e4db207c18cdf12b52
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
